+++ src/pymovements/utils/__init__.py
@@ -33,8 +33,8 @@
     pymovements.utils.paths
 """
 from pymovements.utils import archives  # noqa: F401
-from pymovements.utils import asc2csv  # noqa: F401
 from pymovements.utils import checks  # noqa: F401
 from pymovements.utils import decorators  # noqa: F401
 from pymovements.utils import downloads  # noqa: F401
+from pymovements.utils import parsing  # noqa: F401
 from pymovements.utils import paths  # noqa: F401
